Simplify NRF driver

+++ VCU-APP/Core/Src/Nodes/VCU.c
@@ -284,7 +284,7 @@ uint8_t VCU_CAN_TX_MixedData(hbar_runner_t *runner) {
   TxData.u8[1] = BMS.d.soc;
   TxData.u8[2] = VCU.d.speed; //runner->mode.d.report[HBAR_M_REPORT_RANGE];
   TxData.u8[3] = runner->mode.d.report[HBAR_M_REPORT_AVERAGE];
-  TxData.u32[1] = (uint32_t) VCU.d.gps.dop_h; // VCU.d.odometer;
+  TxData.u32[1] = VCU.d.gps.dop_h; // VCU.d.odometer;
 
   // send message
   return CANBUS_Write(CAND_VCU_SELECT_SET, &TxData, 8);
